+++ pkg/services/object/get/get_test.go
@@ -78,7 +78,7 @@ func (g *testNeoFS) GetNodesForObject(addr oid.Address) ([][]netmap.NodeInfo, []
 	return nodeLists, primaryNums, nil, nil
 }
 
-func (c *testClientCache) get(info client.NodeInfo) (getClient, error) {
+func (c *testClientCache) get(_ context.Context, info client.NodeInfo) (getClient, error) {
 	v, ok := c.clients[info.AddressGroup().String()]
 	if !ok {
 		return nil, errors.New("could not construct client")
@@ -96,7 +96,7 @@ func newTestClient() *testClient {
 	}
 }
 
-func (c *testClient) getObject(exec *execCtx, _ client.NodeInfo) (*object.Object, io.ReadCloser, error) {
+func (c *testClient) getObject(exec *execCtx) (*object.Object, io.ReadCloser, error) {
 	v, ok := c.results[exec.address()]
 	if !ok {
 		var errNotFound apistatus.ObjectNotFound