+++ pkg/services/util/sign.go
@@ -11,6 +11,29 @@ import (
 	protostatus "github.com/nspcc-dev/neofs-sdk-go/proto/status"
 )
 
+// Request is a common interface of API request messages.
+type Request interface {
+	GetMetaHeader() *protosession.RequestMetaHeader
+}
+
+// VersionLE checks whether proto version in request meta header is less or
+// equal than the specified one.
+func VersionLE(req Request, mjr, mnr uint32) bool {
+	ver := req.GetMetaHeader().GetVersion()
+	gotMjr := ver.GetMajor() // NPE-safe
+	return gotMjr < mjr || gotMjr == mjr && ver.GetMinor() <= mnr
+}
+
+// SignResponseIfNeeded checks whether response for the req should be signed. If
+// so, calculated verification header is returned. Otherwise, nil returns.
+func SignResponseIfNeeded[R sdkcrypto.ProtoMessage](signer *ecdsa.PrivateKey, r sdkcrypto.SignedResponse[R], req Request) *protosession.ResponseVerificationHeader {
+	if VersionLE(req, 2, 21) {
+		return SignResponse(signer, r)
+	}
+
+	return nil
+}
+
 func SignResponse[R sdkcrypto.ProtoMessage](signer *ecdsa.PrivateKey, r sdkcrypto.SignedResponse[R]) *protosession.ResponseVerificationHeader {
 	verHeader, err := sdkcrypto.SignResponseWithBuffer(sdkecdsa.Signer(*signer), r, nil)
 	if err != nil {